requirements.txt
Flask
edge-tts
gunicorn
docx
pdfminer.six
openpyxl
pytesseract
pillow